✨ feat(server): Collection service + access + sync → commonMain (native)
