+++ server/src/jvmMain/kotlin/com/calypsan/listenup/server/auth/AuthServiceImpl.kt
@@ -18,8 +18,6 @@ import com.calypsan.listenup.api.dto.auth.SessionId
 import com.calypsan.listenup.api.dto.auth.SessionSummary
 import com.calypsan.listenup.api.dto.auth.User
 import com.calypsan.listenup.api.dto.auth.UserId
-import com.calypsan.listenup.api.dto.auth.UserRole
-import com.calypsan.listenup.api.dto.auth.UserStatus
 import com.calypsan.listenup.api.error.AuthError
 import com.calypsan.listenup.api.result.AppResult
 import com.calypsan.listenup.server.db.UserRoleColumn
@@ -450,24 +448,3 @@ class AuthServiceImpl(
         )
     }
 }
-
-internal fun UserRoleColumn.toContract(): UserRole =
-    when (this) {
-        UserRoleColumn.ROOT -> UserRole.ROOT
-        UserRoleColumn.ADMIN -> UserRole.ADMIN
-        UserRoleColumn.MEMBER -> UserRole.MEMBER
-    }
-
-internal fun UserRole.toColumn(): UserRoleColumn =
-    when (this) {
-        UserRole.ROOT -> UserRoleColumn.ROOT
-        UserRole.ADMIN -> UserRoleColumn.ADMIN
-        UserRole.MEMBER -> UserRoleColumn.MEMBER
-    }
-
-internal fun UserStatusColumn.toContract(): UserStatus =
-    when (this) {
-        UserStatusColumn.ACTIVE -> UserStatus.ACTIVE
-        UserStatusColumn.PENDING_APPROVAL -> UserStatus.PENDING_APPROVAL
-        UserStatusColumn.DENIED -> UserStatus.DENIED
-    }
